Latest Version of Website - restructured folders and tuned user managerment
have problems dealing with the footer of teh website once the PHP return
too much data it doesn't align with the height of the data but stays at
the bottom of the browsers page.
Roles Table is not integrated in the User management and administrators
canupdate the role of a user now as well.