aircraft management refined
If there is only on model assigned to the airfield it is not deleted
from the table but only the model value is set to  'No Aircraft'
The GUI then changes for the user
Case the user adds then a new aircraft no INSERT is executed but an
UPDATE to get the first model into the table.
That way I avoid some error messages and the airfield , even if empty
stays int he drop downs.