dump entire boswar_db
I've made changes to many of the tables, usually just to size of
integers.  This does not change column names nor data in the columns,
though some tables now have extra columns and may have different
indexing.
I dumped the whole database to make it easy to import rather than
importing the changed tables one-at-a-time.

I will use this as the basis for future column updates (which will most
likely be related to mission and campaign statistics).