Cycle completed - create and drop campaign database is now possible through interface
Using the already mentioned interface to create a new campaign works.
I added an additional interface to drop a campaign database.
the list displayed though is limited to hidden or proposed status
(security from user error)
The script also deleted the database user that was bound to the database
during creation - we will have to tweak this with a check if the user is
used in additional campaigns as well and only remove grants on the
deleted database then.