this is the working version - please apply and retry
your error reminded me on an issue I had as well. MySQL proposes to have
the hostname attached to the username and initially I combined
usernameAThostname. This lead to exactly the error you described. i
changed that back and inserted the user without the AT, and created the
DB user with the AThostname in my latest version.
It is strange that you still see this behaviour.
Please retry.