update create_rofwar_campaign_user.sql
Used incorrect comment style and had wrong username and password and
insufficient permissions to undo stats.  But other than those,
everything was right in the earlier version.  It just didn't work.  :)

This one has been tested and verified to not only create the user and
grand permissions but those permissions are sufficient for the parser.
I haven't tested other things.