add create_boswar_hq_user.sql
To create a user for the boswar_db with what may be minimal privileges
for what it needs to do.
Please give it a try and report any issues.

I've also revised connect_db.php to use this user once created.

I've done some quick testing of this user and it is able to create new
campaigns and parse logs.