an administrator now can also change the folder path for the group files
The administrator has to choose the DB he want to store the path to in
the dropdown as he is not connected to any campaign db during user
management and the process would not know what campaign to assign. There
will be an error message dropped if the administrator tries to add a
folder path to a campaign he is not yet assigned to.