revised cam_param.php to use campaign_settings
Note this is in top-level, not in missionplanningPackage for development
and testing.  Several lines at the top will need to be removed to make
it a drop-in replacement for the current cam_param.php.

Also note that the current supply point code is ugly and doesn't take
advantage of the flexibility of the supply.point table format.  This was
forced by the need to make it a drop-in replacement.