first build with an installation guide of the boswar_db
additionally the users and example campaigns are in this build.
In the included word doc the first sanity tests are applied which also
help the user understanding user roles and makes him familiar with the
interface.
I covered 90% of the user management so far.
Please review and improve if required.
Please alos test if the process works as I did not test it yet. The
databases are from todays valid versions and the Website has all
published php files in it.
Enjoy testing :)