update parser function coreOUTPUT
It now identifies more of the indirect kills and properly assigns to the
appropriate attacker.  It also eliminates what should be among the final
of the rof-specific references in the parser code, does more scoring,
and now uses the same query structure for all inserts, providing an
opportunity to greatly simplfy the code... but that is a later step...
for after today's mission.