update parser functions coreOUTPUT and processPLAYERNAME
These modifications permit the resolution of the last  remaining
unattributed delayed kills (other than accidents, etc.).  To discourage
kill stealing, credit is given to the pilot who gets the lethal hit on a
plane, for the plane and its pilot and gunners if they are subsequently
killed (unless perhaps they are directly killed by another pilot... but
that is very rare).  We are talking about kill credits, not points.
There are no points to be gained for pilot or gunner kills.