update rof_object_properties
Merging in Stenka's 'vehicles' table - which is the only fully populated
example of his tables.  I'll do the others once they are populated.  The
only column I did not import was game_name which is redundant for
rof_object_properties.  :)  I did note that the "Models" column matches
the existing "object_type" column in a case-insensitive manner.  The
cases do differ between the Log and the Mission Builder.