update CampaignMgmtSetup
Show defaults for the min and max x and z settings.  Put the code in an
include (getMaxMinXZ) to keep in style with Myata's habits.  :)

I also emphasized that we need the max and min values rather than two
corners in case of a diamond-shaped combat area.  We are making a
rectangle that includes the actual combat area here.