add a .htaccess file to set upload_max_filesize
My php.ini file had a default file upload limit of 2 MB, rather than
change it in php.ini I've added this .htaccess file that raises it to 4
MB.  This might be useful for others who have a low default setting.  It
should be the same as our uploadFile.php limit.