create and drop campaign changes
I added a statement that assigns the admin creating the campaign to the
campaign_users table in the boswar_db. That way we have a campaign
administrator role from the beginning and the admin can easily store a
groupFile Path to this campaign once he connects to it (There is now a
revised UserManagement similar to the commander User management
availabel).
The initial role is set neutral and this will later be used to display
both sides values.

The Drop campaign statement was enhanced to include all assigned users
from the campaign_users table as well.
It still need revision to use the connected db as filter instead of the
drop down db.
This way the administrator is only able to change the path for the
actually connected campaing which we now can be sure he is assigned to.