update CampaignMgmtCreateConfirm.php
convert remaining mysqli calls to object oriented style, make a few
capitalization corrections and avoid a warning from strict checking
about no next_result() when come to the end of the multi statement.
(Only seen when not forwarding the page.)  Also added error reporting
for the multi_query... it is useful when not forwarding the page.