User mangement is now having an additional filter
As we were talking about a campaign administrative role this is the
first step towards this.
An administrator has now 2 different user management consoles.

1:  The usual one displayed if he is not connected to any campaign.
In here the admin can manipulate all users but is no longer able to
change his group file path.

2:  Once connected to any campaign the admin sees all users, regardless
if they are already promoted to commanders or not, that are already
connected with the actual campaign. The admin can now change the role
and coalition of that user.

Additionally he can change his group file path for the selected
campaign.

Connected to a campaign the administrator is not able to assign any user
to any other campaign!