restoring campaign creation and configuration to
to the state  it was in about 13 hours ago when everything was working.
Please install campaign_settings.sql and update everything and try it
out, following along in mission editor and doing the configuration as
far as it goes.

We need to get a new campaign set up and configured before we can deal
with planning missions!