object management should be working now 100%
There is a small issue with I find no solution for. If you have the POST
variable debugging turned on then the loacation function is not working
so I disabled it for now.
There is no problem enabling it but my browser throws an error on the
page CampaignMgmtObjectsConfirm.php