update campaign configuration/setup/upload/import
It now works smoothly again, with all important references to $abbrv
back in place.

You should now be able to create a campaign and get it configured
through the update countries & coalitions and max and min X and Z.

Any suggestions for improvement in the instructions so far will be
gratefully accepted.