update CampaignMgmtImportConfirm.php
Finally got the .Group file to delete.  Problem was somewhere within
template_to_Airfield.php.  I turned it into a function to isolate the
problem and the problem went away... an advantage of functions over
includes: they are less likely to alter things you don't want altered.
Anyway it is a relief!