Airfield Management implemented
I hope I caught all bugs - this is as far more complicated as expected
but this solution should do the trick. I did some tests but we will have
to check if the process with the airfields works fine as well.

I do not do updates to the country in case coalition changes - I hope
this is not a requirement.

I removed the coalition from the airfields_models table as it is not
required and would be redundant.

Please update your tables accordingly.