eliminating the camp-airfields.Group step
Thanks, Stenka.

Also correcting a bunch of places where $dbc was used inappropriately.

Once connected to a campaign, $dbc should only appear at the top and
bottom of a file... to read the info needed to connect to the campaign
and to close the connection.  Everything else should be done with
$camp_link.