updating columns.sql
Restoring my version with supplypoint as tinyint(1) for 255 supply
points max per side and with NULL for columns that permit it.  The
latter is for two reasons... one is for better debugging, and the other
is to avoid problems like those Myata had with null bridge spans when
they were default 0.