progress on static groups
NOTE DB UPDATE to static.sql.  Bumped db version again.

Now each member of a static group is written to a line in the static
table.

Next on the agenda... write these out to a group file.

I tested with a one of each Entente and one of each Central Powers
group, and both were written to 'static' with no errors so I think this
is reasonably solid.  Be sure to get fully updated on the DB changes,
though.  It won't work without them.