update three parser functions
outputFATES is the important one... I missed it when converting
functions to the single-value-return style.  outputTURRETGUNNER adds a
debugging line, and outputWHOSEGUNNER adds a missing semicolon to make
it work.

I also had to add two turretgunners to rof_object_properties... and edit
a Sopwith Strutter, but I'll load those changes later.