improved aircraft model manipulation screen
I changed the button texts, added a new filter if a user inserts 0 a new
quantity, the model dissapears from the list, the "Select Aircraft
Model..." dropdown now is bound to a function that lists only the
available models and excludes the ones already assigned to this field.
That way I was able to make the user interface easier to understand.