update bos_stalingrad_locations.sql
Airfields (and their neighboring villages... presumed to be the same
name) in the SW quadrant of the Stalingrad map.  The numbers give the
approximate center of the runway.  If there are multiple runways, there
is an entry for each and the airfield is coded as a '20'.  The town of
Abganerovo was too extensive to be included in a circle with a radius of
750 m, so I used three of them to get most of it.