BosWar buttons
I would say this is the button layout of the BosWar interface. Now I
like it. As long as I have no access to the mission builder and I cannot
take appropriate screenshots , I cannot go ahead from here as the images
on the web are rare and do not achive my expectations.