update bos_stalingrad_locations
100% done for airfields, airfield villages and Stalingrad (226 entries
in all).  There are 90 airfields, of which 72 have one runway and 18
have two.  I have also added entries for North, Central and South
Stalingrad, as well as the entries for the villages closest to most
airfields (and presumed to carry the same name...).  I'll add no more
cities or towns/villages unless and until they are named on the
in-flight navigation map.  Otherwise they are useless.